requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-ipware==3.0.2
django-modeladmin-reorder==0.3.1
django-modeltranslation==0.17.3
django-rest-framework==0.1.0
django-sms==0.5.0
djangorestframework==3.12.4
fake-useragent==0.1.11
idna==2.10
lxml==4.6.3
Pillow==8.2.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==1.7.1
python-geoip==1.2
python-geoip-geolite2==2015.303
python-geoip-python3==1.3
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
twilio==6.62.0
urllib3==1.26.5
python-social-auth==0.3.6
python3-openid==3.2.0
requests-oauthlib==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0